test(e2e): split tunnel workflow checks
